Prepare to resolve remote type metadata documents into credential schemes
