+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/data/ConstantIndex.kt
@@ -221,82 +221,3 @@ interface VcJwtCredentialScheme : CredentialScheme {
         get() = listOf(PLAIN_JWT)
 }
 
-
-private fun SdJwtTypeMetadata.extractRepresentation() = when (vckExtensions?.format) {
-    CredentialFormatEnum.JWT_VC -> PLAIN_JWT
-    CredentialFormatEnum.DC_SD_JWT -> SD_JWT
-    CredentialFormatEnum.MSO_MDOC -> ISO_MDOC
-    else -> null
-} ?: SD_JWT
-
-
-fun SdJwtTypeMetadata.toCredentialScheme() = when (extractRepresentation()) {
-    PLAIN_JWT -> ExtractedVcJwtCredentialScheme(
-        schemaUri = "TODO",
-        vcType = vckExtensions?.vcType ?: vct.string,
-        claimDescriptions = claims?.map { it.toClaimDescription() }?.toSet() ?: setOf()
-    )
-
-    SD_JWT -> ExtractedSdJwtCredentialScheme(
-        schemaUri = "TODO",
-        sdJwtType = vct.string,
-        claimDescriptions = claims?.map { it.toClaimDescription() }?.toSet() ?: setOf()
-    )
-
-    ISO_MDOC -> ExtractedIsoMdocCredentialScheme(
-        schemaUri = "TODO",
-        isoDocType = vckExtensions?.isoDocType ?: vct.string,
-        isoNamespace = vckExtensions?.isoNamespace ?: vct.string,
-        claimDescriptions = claims?.map { it.toClaimDescription() }?.toSet() ?: setOf()
-    )
-}
-
-private fun SdJwtTypeMetadataClaimInformation.toClaimDescription(): ClaimDescription = ClaimDescription(
-    path = OpenId4VciClaimsPathPointer(path.map {
-        it.toSegment()
-    }.toNonEmptyList()),
-    mandatory = isMandatory,
-    display = display?.map {
-        DisplayProperties(
-            locale = it.locale.string,
-            name = it.label,
-            description = it.description,
-        )
-    }?.toSet()
-)
-
-private fun SdJwtTypeMetadataClaimInformationPathSegment?.toSegment(): OpenId4VciClaimsPathPointerSegment? =
-    when (this) {
-        is SdJwtTypeMetadataClaimInformationPathSegmentIndex ->
-            OpenId4VciClaimsPathPointerSegmentIndex(this.ulong.safeToUint())
-
-        is SdJwtTypeMetadataClaimInformationPathSegmentName ->
-            OpenId4VciClaimsPathPointerSegmentString(this.string)
-
-        null -> null
-    }
-
-private fun ULong.safeToUint(): UInt = also {
-    if (it > UInt.MAX_VALUE) {
-        throw UnsupportedOperationException("This implementation only supports claims path pointer indices up to ${UInt.MAX_VALUE}, but got $it")
-    }
-}.toUInt()
-
-data class ExtractedVcJwtCredentialScheme(
-    override val schemaUri: String,
-    override val vcType: String,
-    override val claimDescriptions: Set<ClaimDescription>
-) : VcJwtCredentialScheme
-
-data class ExtractedSdJwtCredentialScheme(
-    override val schemaUri: String,
-    override val sdJwtType: String,
-    override val claimDescriptions: Set<ClaimDescription>
-) : SdJwtCredentialScheme
-
-data class ExtractedIsoMdocCredentialScheme(
-    override val schemaUri: String,
-    override val isoDocType: String,
-    override val isoNamespace: String,
-    override val claimDescriptions: Set<ClaimDescription>
-) : IsoMdocCredentialScheme
